webstorage/storage-set-value-enumerate.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>WebStorage Test: Storage - set value with enumerate value</title>
<link rel="author" title="Intel" href="http://www.intel.com" >
<link rel="help" href="https://w3c.github.io/webstorage/#the-storageevent-interface" >
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  var store_list = new Array("key0", "value0", "key1", "value1",
                             "key2", "value2", "key3", "value3",
                             "key4", "value4");
  ["localStorage", "sessionStorage"].forEach(function(name) {
    test(function () {
      var storage = window[name];
      storage.clear();

      for (var i=0; i<store_list.length; i++) {
        if (i == 0)
          continue;
        if((i%2) == 1) {
          storage.setItem(store_list[i-1], store_list[i]);
          if(!(storage.length == parseInt(i/2+1)))
            assert_unreached("length is not correct");
        }
      }

      for (var i=0; i<store_list.length; i++) {
        if((i%2) == 0) {
          var key = store_list[i];
          var value = storage.getItem(key);
          assert_equals(value, store_list[i+1]);
        }
      }

    }, "Check if " + name + "set values with " + name + ".key = value");
  });
</script>
